Fix wall dectection and refactor
So, there was some sort of refactor half-done.
I don't really remember anymore.
But this code was using var binding fuckery and wall detection was half-assed or half-written in a confusing code.
So now wall detection is fixed, some rubbish deleted, and some improved.
Next step is to use ES6 style everywhere.